requirements.txt
click==7.1.2
Flask==1.1.2
geojson==2.5.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
overpy==0.4
Pillow==8.0.1
urllib3==1.26.2
Werkzeug==1.0.1
